fix(IN-320): use client 분리 복원
SEO 적용은 app 폴더 안에서만 진행되므로 src/pages안의 페이지들은 use client 상태를 유지해도 상관이 없기에 파일을 복원함
