+++ src/pages/main/ui/ChannelProfilePage.tsx
@@ -1,5 +1,24 @@
-import { ChannelProfileClient } from './ChannelProfileClient'
+'use client'
+
+import { ChannelProfileSection } from '@/widgets/main/channelProfile'
+import { TrendingVideosSection } from '@/widgets/main/trendingVideos'
+import { TrendMagazineSection } from '@/widgets/main/trendMagazine'
+import { useAuth } from '@/features/auth'
 
 export function ChannelProfilePage() {
-  return <ChannelProfileClient />
+  const { user } = useAuth()
+  const channelId = user?.userChannelDetails?.youtubeChannelId as string
+
+  return (
+    <div className='flex flex-col divide-y'>
+      {/* 프로필 카드 */}
+      <ChannelProfileSection channelId={channelId} isExpanded={false} />
+
+      {/* 콘텐츠 */}
+      <div className='felx h-fit w-full flex-col px-24 pt-40 pb-96'>
+        <TrendingVideosSection channelId={channelId} />
+        <TrendMagazineSection channelId={channelId} />
+      </div>
+    </div>
+  )
 }
